Don't break up inline math during response streaming
Fixes #258571
