+++ src/vs/workbench/contrib/chat/common/chatWordCounter.ts
@@ -3,6 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import * as markedKatexExtension from '../../markdown/common/markedKatexExtension.js';
+
 export interface IWordCountResult {
 	value: string;
 	returnedWordCount: number;
@@ -43,11 +45,14 @@ const linkPattern =
 export function getNWords(str: string, numWordsToCount: number): IWordCountResult {
 	// This regex matches each word and skips over whitespace and separators. A word is:
 	// A markdown link
+	// Inline math
 	// One chinese character
 	// One or more + - =, handled so that code like "a=1+2-3" is broken up better
 	// One or more characters that aren't whitepace or any of the above
 	const backtick = '`';
-	const allWordMatches = Array.from(str.matchAll(new RegExp(linkPattern + r`|\p{sc=Han}|=+|\++|-+|[^\s\|\p{sc=Han}|=|\+|\-|${backtick}]+`, 'gu')));
+
+	const wordRegExp = new RegExp('(?:' + linkPattern + ')|(?:' + markedKatexExtension.mathInlineRegExp.source + r`)|\p{sc=Han}|=+|\++|-+|[^\s\|\p{sc=Han}|=|\+|\-|${backtick}]+`, 'gu');
+	const allWordMatches = Array.from(str.matchAll(wordRegExp));
 
 	const targetWords = allWordMatches.slice(0, numWordsToCount);
