+++ src/vs/workbench/contrib/chat/test/common/chatWordCounter.test.ts
@@ -102,6 +102,14 @@ suite('ChatWordCounter', () => {
 
 			cases.forEach(([str, nWords, result]) => doTest(str, nWords, result));
 		});
-	});
 
+		test(`Inline math shouldn't be broken up`, () => {
+			const cases: [string, number, string][] = [
+				['a $x + y$ b', 3, 'a $x + y$ b'],
+				['a $\\frac{1}{2} + \\sqrt{x^2 + y^2}$ b', 3, 'a $\\frac{1}{2} + \\sqrt{x^2 + y^2}$ b'],
+			];
+
+			cases.forEach(([str, nWords, result]) => doTest(str, nWords, result));
+		});
+	});
 });
